Added logging information to migration script, also made minor changes to it

+++ migration/Migrate.py
@@ -9,6 +9,8 @@ from libcloud.storage.types import InvalidContainerNameError
 import random
 import string
 import paramiko
+from glanceclient import Client
+import logging
 
 
 class Migrate:
@@ -24,46 +26,65 @@ class Migrate:
         self.__location = None
         self.__vols = []
 
+        self.logger = logging.getLogger(__name__)
+
         self.aws_prov = AWSProvider()
         self.os_prov = OpenStackProvider()
 
     def deploy_S3(self):
+        """
+        Method to create a randomly identifiable S3 bucket.
+        :return: None
+        """
         attempts = 0
         while True:
             try:
+                self.logger.debug("Attempt number {} to create S3".format(attempts))
                 attempts += 1
                 randomName = ''.join([random.choice(string.ascii_letters + string.digits) for n in range(32)])
+                self.logger.debug("Random name for S3 {}".format(randomName))
                 container = self.aws_prov.create_container(randomName)
+                self.logger.info("S3 Bucket {} created".format(randomName))
                 return container
-            except InvalidContainerNameError:
+            except InvalidContainerNameError as e:
                 if attempts > 3:
+                    self.logger.exception("Failed to create an S3")
                     raise InvalidContainerNameError
                 continue
 
     def connect_to_node(self):
+        """
+        Function to connect to the node which is to be migrated
+        :return: None
+        """
         ip = self.node.public_ips[0]
+        self.logger.debug("IP Address of node to be migrated: {}".format(ip))
         self.ssh = paramiko.SSHClient()
 
         self.ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
 
         self.ssh.connect(ip, username='ubuntu', password='',
-                    key_filename=self.node_key_file_loc)
+                         key_filename=self.node_key_file_loc)
 
-    def copy_image(self):
-        # TODO: 1) Find the volume size and type ... type may be less important Done
-        # TODO: 2) Create a volume of the same size on AWS
-        # TODO: 3) Attach that volume to the instance
-        # TODO: 4) Partition, format and mount the disk
-        # TODO: 5) Copy the disk to an image file
+        self.logger.info("Connection created to node to be migrated.")
+
+    def copy_image_local(self):
+        """
+        Function to copy the root volume of a node to a newly created node of the same size
+        :return: None
+        """
+
         block_devices = self.node.extra["block_device_mapping"]
         self.__location = self.node.extra["availability"]
+        self.logger.debug("Location of aws node to be migrated {}".format(self.__location))
 
         for vol in block_devices:
             volume_id = vol["ebs"]["volume_id"]
             self.__vols.append(self.aws_prov.get_volume(volume_id))
+            self.logger.debug("Volume attached to the migrated instance {}".format(volume_id))
+
         migration_vols = []
         vol_count = 0
-
         mounts = "abcdefghijklmnopqrstuvwxyz"
         for vol in self.__vols:
             migrate_vol = self.aws_prov.create_volume("migration_vol", vol.size, location=self.__location)
@@ -71,50 +92,44 @@ class Migrate:
 
             device_name = "/dev/xvd{}".format(mounts[vol_count])
             mount_point = "tmp/disk{}".format(vol_count)
+            self.logger.info("Device attached: {}, Mount Point: {}, Size: {}".format(device_name,
+                                                                                     mount_point,
+                                                                                     vol.size))
 
             success = self.aws_prov.attach_volume(self.node, migrate_vol, device=device_name)
 
-            stdin, stdout, stderr = self.ssh.exec_command("lsblk")
-            for line in stdout.readlines():
-                print(line.split())
-            for line in stderr.readlines():
-                print(line, end="")
+            if not success:
+                self.logger.critical("Volume could not be attached")
+                return
+
+            self.logger.info("Volume attached")
 
             stdin, stdout, stderr = self.ssh.exec_command("sudo mkfs -t ext4 {}".format(device_name))
-            for line in stdout.readlines():
-                print(line, end="")
-            for line in stderr.readlines():
-                print(line, end="")
+            self.logger.info(stdout.readlines)
+            self.logger.warning(stderr.readlines)
 
             stdin, stdout, stderr = self.ssh.exec_command("sudo mkdir {}".format(mount_point))
-            for line in stdout.readlines():
-                print(line, end="")
-            for line in stderr.readlines():
-                print(line, end="")
+            self.logger.info(stdout.readlines)
+            self.logger.warning(stderr.readlines)
 
             stdin, stdout, stderr = self.ssh.exec_command("sudo mount {} {}".format(device_name, mount_point))
-            for line in stdout.readlines():
-                print(line, end="")
-            for line in stderr.readlines():
-                print(line, end="")
+            self.logger.info(stdout.readlines)
+            self.logger.warning(stderr.readlines)
 
-            stdin, stdout, stderr = self.ssh.exec_command("sudo dd if={} of={}/disk.img".format(device_name, mount_point))
-            for line in stdout.readlines():
-                print(line, end="")
-            for line in stderr.readlines():
-                print(line, end="")
+            stdin, stdout, stderr = self.ssh.exec_command(
+                "sudo dd if={} of={}/disk.img".format(device_name, mount_point))
+            self.logger.info(stdout.readlines)
+            self.logger.warning(stderr.readlines)
 
             vol_count += 1
+            self.logger.info("Volume image successfully copied.")
 
-    def transfer_image(self, container):
-        # TODO: 0) Possibly create the s3 bucket
-        # TODO: 1) Add AWS credentials to the remote instance (the being copied) or else upload the thing using libcloud
-        # TODO: 2) Copy the image file to the s3 bucket
-        # TODO: 3) Unmount and delete the volume
-        # TODO: 4) Copy image from s3 to local
-        # Create the container if it doesn't exist
-        if container is None:
-            container = self.deploy_S3()
+    def transfer_image_to_s3(self, container):
+        """
+        Method to transfer the image from the new volume to the provided S3
+        :param container: S3 Bucket we are to transfer the img file to.
+        :return: None
+        """
         bucket_name = container.name
 
         # Add aws credentials to the instance to be copied.
@@ -122,28 +137,42 @@ class Migrate:
         stdin, stdout, stderr = self.ssh.exec_command("""aws configure set aws_access_key_id {};
         aws configure set aws_secret_access_key {};
         aws configure set default.region {}""").format(aws_access_id, aws_secret_key, self.__location)
-        for line in stdout.readlines():
-            print(line, end="")
-        for line in stderr.readlines():
-            print(line, end="")
+        self.logger.info(stdout.readlines)
+        self.logger.warning(stderr.readlines)
 
         # Copy each volume to the s3
         vol_count = 0
         for vol in self.__vols:
-            self.ssh.exec_command("aws s3 cp/tmp/disk{} s3://{}/".format(i, bucket_name))
-            for line in stdout.readlines():
-                print(line, end="")
-            for line in stderr.readlines():
-                print(line, end="")
+            self.ssh.exec_command("aws s3 cp /tmp/disk{}.img s3://{}/".format(vol_count, bucket_name))
+            self.logger.info(stdout.readlines)
+            self.logger.warning(stderr.readlines)
 
             # Once copied to the s3 dettach and destroy the volume
             self.aws_prov.dettach_volume(self.node, vol)
             self.aws_prov.destroy_volume(vol)
+            self.logger.info("Volume removed and destroyed")
 
+            vol_count += 1
+        return vol_count
 
-    def create_new_node(self):
-        # TODO: 0) Optionally destroy the instance itself
-        # TODO: 1) Create image from the raw disk image
-        # TODO: 2) Create node from this image
-        # extra: Check if we need a means of translating the sizes and such
+    def pull_image(self, container, file_name, dest):
+        """
+        Dowload the image file to be turned into a node
+        :param container: container to pull the disk from.
+        :param file_name: file to be pulled i.e. the image file.
+        :param dest: Location where image file should be placed.
+        :return: None
+        """
+        # Download the image file
+        success = self.aws_prov.download_from_container(container, file_name, dest)
+        if not success:
+            self.logger.error("Failed to download the img file from s3")
+            return
+
+    def create_image(self, image_file_loc):
         raise NotImplementedError
+
+
+
+    def create_node(self):
+        print("Glance part")
